Moved distributed lock tests
